Reorganize folder structure for better naming consistency
- Rename schema-request to request-schema
- Rename schema-response to response-schema
- Rename validator-request to request-validator
- Rename validator-response to response-validator
- Update all import paths across the codebase (32 files)
- Maintain intuitive naming: what-type pattern (request-schema, response-validator)
- All changes compile successfully
