+++ packages/kori/src/kori/kori-creator.ts
@@ -2,8 +2,8 @@ import { type KoriEnvironment, type KoriRequest, type KoriResponse } from '../co
 import { createKoriLoggerFactory, createInstanceLogger } from '../logging/index.js';
 import { type KoriLoggerFactoryOptions, type KoriLoggerFactory } from '../logging/index.js';
 import { createHonoRouter, type KoriRouter } from '../router/index.js';
-import { type KoriRequestValidatorDefault } from '../validator-request/index.js';
-import { type KoriResponseValidatorDefault } from '../validator-response/index.js';
+import { type KoriRequestValidatorDefault } from '../request-validator/index.js';
+import { type KoriResponseValidatorDefault } from '../response-validator/index.js';
 
 import { createKoriInternal, type KoriInternalShared } from './kori-internal.js';
 import { type Kori } from './kori.js';
