+++ packages/kori/src/kori/kori-internal.ts
@@ -5,10 +5,10 @@ import { getMethodString } from '../http/index.js';
 import { type KoriLogger } from '../logging/index.js';
 import { type KoriPlugin } from '../plugin/index.js';
 import { type KoriRouter, type KoriRouterHandler } from '../router/index.js';
-import { type KoriRequestSchemaDefault } from '../schema-request/index.js';
-import { type KoriResponseSchemaDefault } from '../schema-response/index.js';
-import { type KoriRequestValidatorDefault } from '../validator-request/index.js';
-import { type KoriResponseValidatorDefault } from '../validator-response/index.js';
+import { type KoriRequestSchemaDefault } from '../request-schema/index.js';
+import { type KoriResponseSchemaDefault } from '../response-schema/index.js';
+import { type KoriRequestValidatorDefault } from '../request-validator/index.js';
+import { type KoriResponseValidatorDefault } from '../response-validator/index.js';
 
 import { createFetchHandler } from './fetch-handler-factory.js';
 import { type Kori, type KoriRouteDefinition } from './kori.js';
