+++ packages/kori/src/kori/kori.ts
@@ -3,10 +3,10 @@ import { type KoriFetchHandler } from '../fetch-handler/index.js';
 import { type KoriOnErrorHook, type KoriOnRequestHook, type KoriOnStartHook } from '../hook/index.js';
 import { type KoriLogger } from '../logging/index.js';
 import { type KoriPlugin } from '../plugin/index.js';
-import { type KoriRequestSchemaDefault } from '../schema-request/index.js';
-import { type KoriResponseSchemaDefault } from '../schema-response/index.js';
-import { type KoriRequestValidatorDefault } from '../validator-request/index.js';
-import { type KoriResponseValidatorDefault } from '../validator-response/index.js';
+import { type KoriRequestSchemaDefault } from '../request-schema/index.js';
+import { type KoriResponseSchemaDefault } from '../response-schema/index.js';
+import { type KoriRequestValidatorDefault } from '../request-validator/index.js';
+import { type KoriResponseValidatorDefault } from '../response-validator/index.js';
 
 import { type RequestProviderCompatibility, type ResponseProviderCompatibility } from './route-options.js';
 import {
