+++ packages/kori/src/request-validator/validate-body.ts
@@ -1,7 +1,7 @@
 import { type KoriRequest } from '../context/index.js';
 import { ContentType } from '../http/index.js';
 import { type KoriSchemaDefault, isKoriSchema } from '../schema/index.js';
-import { type KoriRequestSchemaContentBodyDefault, type KoriRequestSchemaDefault } from '../schema-request/index.js';
+import { type KoriRequestSchemaContentBodyDefault, type KoriRequestSchemaDefault } from '../request-schema/index.js';
 import { ok, err, type KoriResult } from '../util/index.js';
 
 import { type KoriBodyValidationError } from './error.js';
