+++ packages/kori/src/request-validator/validate-headers.ts
@@ -1,24 +1,24 @@
 import { type KoriRequest } from '../context/index.js';
-import { type KoriRequestSchemaDefault } from '../schema-request/index.js';
+import { type KoriRequestSchemaDefault } from '../request-schema/index.js';
 import { ok, err, type KoriResult } from '../util/index.js';
 
 import { type KoriFieldValidationError } from './error.js';
 import { type KoriRequestValidatorDefault } from './request-validator.js';
 
-export async function validateRequestQueries({
+export async function validateRequestHeaders({
   validator,
   schema,
   req,
 }: {
   validator: KoriRequestValidatorDefault;
-  schema: KoriRequestSchemaDefault['queries'];
+  schema: KoriRequestSchemaDefault['headers'];
   req: KoriRequest;
 }): Promise<KoriResult<unknown, KoriFieldValidationError<unknown>>> {
   if (!schema) {
     return ok(undefined);
   }
 
-  const result = await validator.validateQueries({ schema, queries: req.queryParams() });
+  const result = await validator.validateHeaders({ schema, headers: req.headers() });
   if (result.ok) {
     return result;
   }
