+++ packages/kori/src/request-validator/validate-queries.ts
@@ -1,24 +1,24 @@
 import { type KoriRequest } from '../context/index.js';
-import { type KoriRequestSchemaDefault } from '../schema-request/index.js';
+import { type KoriRequestSchemaDefault } from '../request-schema/index.js';
 import { ok, err, type KoriResult } from '../util/index.js';
 
 import { type KoriFieldValidationError } from './error.js';
 import { type KoriRequestValidatorDefault } from './request-validator.js';
 
-export async function validateRequestParams({
+export async function validateRequestQueries({
   validator,
   schema,
   req,
 }: {
   validator: KoriRequestValidatorDefault;
-  schema: KoriRequestSchemaDefault['params'];
+  schema: KoriRequestSchemaDefault['queries'];
   req: KoriRequest;
 }): Promise<KoriResult<unknown, KoriFieldValidationError<unknown>>> {
   if (!schema) {
     return ok(undefined);
   }
 
-  const result = await validator.validateParams({ schema, params: req.pathParams() });
+  const result = await validator.validateQueries({ schema, queries: req.queryParams() });
   if (result.ok) {
     return result;
   }
