+++ packages/kori/src/response-schema/entry-content.ts
