+++ packages/kori/src/response-schema/entry-simple.ts
