+++ packages/kori/src/response-validator/infer.ts
